% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_set.R
\name{ma_set}
\alias{ma_set}
\title{Example Pass-By-Reference Function}
\usage{
ma_set(x, i, rp)
}
\arguments{
\item{x}{a \bold{mutatomic} vector (also works on arrays). \cr}

\item{i}{a vector of strictly positive numbers, providing vector indices. \cr}

\item{rp}{the replacement value. \cr
Must be of the same type as \code{x}, and the same length or lenght of 1. \cr \cr}
}
\value{
Returns: VOID. This method modifies the object by reference. \cr
Do not use assignments like \code{x <- ma_set(x, ...)}. \cr
Since this function returns void, you'll just get \code{NULL}. \cr \cr
}
\description{
The \code{ma_set()} function modifies a subset of a \link{mutatomic} object by reference. \cr
It is an example function, used to illustrate various examples in the help pages. \cr
For full pass-by-reference functionality,
please see - for example - the 'squarebrackets' package. \cr \cr
}
\examples{
x <- as.mutatomic(1:16)
ma_set(x, i = 1:6, rp = 8.5) # 8.5 coerced to 8, because `x` is of type `integer`
print(x)
}
