% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mutatomic.R
\name{mutatomic_methods}
\alias{mutatomic_methods}
\alias{as.logical.mutatomic}
\alias{as.integer.mutatomic}
\alias{as.double.mutatomic}
\alias{as.complex.mutatomic}
\alias{as.character.mutatomic}
\alias{as.raw.mutatomic}
\alias{c.mutatomic}
\alias{[.mutatomic}
\alias{[<-.mutatomic}
\alias{format.mutatomic}
\alias{print.mutatomic}
\title{'mutatomic' Method Extension of Generic Functions}
\usage{
\method{as.logical}{mutatomic}(x, ...)

\method{as.integer}{mutatomic}(x, ...)

\method{as.double}{mutatomic}(x, ...)

\method{as.complex}{mutatomic}(x, ...)

\method{as.character}{mutatomic}(x, ...)

\method{as.raw}{mutatomic}(x, ...)

\method{c}{mutatomic}(..., use.names = TRUE)

\method{[}{mutatomic}(x, ...)

\method{[}{mutatomic}(x, ...) <- value

\method{format}{mutatomic}(x, ...)

\method{print}{mutatomic}(x, ...)
}
\arguments{
\item{x}{an atomic object.}

\item{...}{method dependent arguments.}

\item{use.names}{Boolean, indicating if \link[base]{names} should be preserved.}

\item{value}{see \link[base]{Extract}.}
}
\value{
Returns, modifies, or prints a \code{mutatomic} object. \cr
\cr
}
\description{
This page lists the methods defined for the \link{mutatomic} class. \cr
For the package overview, see \link{mutatomic_help}. \cr
}
\examples{

x <- mutatomic(
  1:20, dim = c(5, 4), dimnames = list(letters[1:5], letters[1:4])
)
x

x <- matrix(1:10, ncol = 2)
x <- as.mutatomic(x)
is.mutatomic(x)
print(x)
x[, 1]
x[] <- as.double(x)
print(x)
is.mutatomic(x)
}
