% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer.R
\name{.internal_set_ma}
\alias{.internal_set_ma}
\title{Exposed functions for Package developers}
\usage{
.internal_set_ma(x)
}
\arguments{
\item{x}{atomic object}
}
\value{
Returns: VOID. This function modifies the object by reference. \cr
Do NOT use assignment like \code{x <- .internal_set_ma(x, ...)}. \cr
Since this function returns void, you'll just get \code{NULL}. \cr \cr
}
\description{
Functions for developers. \cr
The functions in this list should NOT be called by regular users. \cr
They are only to be used inside packages. \cr
\cr
}
\details{
\code{.internal_set_ma()}
sets an object to class 'mutatomic' by reference. \cr
}
\examples{


testfun1 <- function(x) {
  .internal_set_ma(x)
}


x <- 1:10
is.mutatomic(x)

testfun1(x)
is.mutatomic(x)
print(x)


}
